#2 by order process
it appears that data orders do not differ in the data objects
downloaded. Looks like the selection of variables refers to the intended
use only, not what is actually being downloaded.

Some problesm with eduation, drop the RAND harmonized effort, and focus,
as previously ont he individual waves. This approach appears to be more
general and self-reliant.